"Fix" error when creating users.
"Fixed" a unique constraint error caused by multiple users having the same username (which would be no username at all, since the e-mail is being used for authentication). This was done by setting the user's username to their email on save. This is a temporary fix which will be addressed later, and the only reason it is being used as a fix for now is because the project, in its current, initial state, is due today.